requirements.txt
streamlit
openllm
langchain
PyMuPDF
beautifulsoup4
requests
python-dotenv
langchain-community
selenium